Move hasNamespace to Element namespace
