+++ packages/alfa-rules/src/common/predicate/is-document-element.ts
@@ -1,14 +1,10 @@
-import {
-  Element,
-  hasName,
-  hasNamespace,
-  Namespace,
-} from "@siteimprove/alfa-dom";
+import { Element, hasName, Namespace } from "@siteimprove/alfa-dom";
 import { Predicate } from "@siteimprove/alfa-predicate";
 
+const { isElement, hasNamespace } = Element;
 const { and } = Predicate;
 
 export const isDocumentElement = and(
-  Element.isElement,
+  isElement,
   and(hasName("html"), hasNamespace(Namespace.HTML))
 );
