+++ packages/alfa-rules/src/sia-r47/rule.ts
@@ -2,7 +2,6 @@ import { Rule } from "@siteimprove/alfa-act";
 import {
   Element,
   hasName,
-  hasNamespace,
   Namespace,
 } from "@siteimprove/alfa-dom";
 import { clamp } from "@siteimprove/alfa-math";
@@ -15,6 +14,7 @@ import { expectation } from "../common/expectation";
 
 import { hasAttribute } from "../common/predicate/has-attribute";
 
+const { isElement, hasNamespace } = Element;
 const { and, equals } = Predicate;
 
 export default Rule.Atomic.of<Page, Element>({
@@ -26,7 +26,7 @@ export default Rule.Atomic.of<Page, Element>({
           .descendants()
           .filter(
             and(
-              Element.isElement,
+              isElement,
               and(
                 hasNamespace(Namespace.HTML),
                 hasName("meta"),
