+++ packages/alfa-table/src/row.ts
@@ -1,5 +1,5 @@
 import { Comparable } from "@siteimprove/alfa-comparable";
-import { Element, isElementByName } from "@siteimprove/alfa-dom";
+import { Element } from "@siteimprove/alfa-dom";
 import { Equatable } from "@siteimprove/alfa-equatable";
 import { Serializable } from "@siteimprove/alfa-json";
 import { Err, Ok, Result } from "@siteimprove/alfa-result";
@@ -7,6 +7,7 @@ import { Err, Ok, Result } from "@siteimprove/alfa-result";
 import * as json from "@siteimprove/alfa-json";
 
 import { Cell } from "./cell";
+import {  isElementByName } from "./helpers";
 
 const { compare } = Comparable;
