+++ packages/alfa-table/src/table.ts
@@ -1,5 +1,5 @@
 import { Comparable } from "@siteimprove/alfa-comparable";
-import { Element, isElementByName, Node } from "@siteimprove/alfa-dom";
+import { Element } from "@siteimprove/alfa-dom";
 import { Equatable } from "@siteimprove/alfa-equatable";
 import { Iterable } from "@siteimprove/alfa-iterable";
 import { Serializable } from "@siteimprove/alfa-json";
@@ -9,6 +9,7 @@ import * as json from "@siteimprove/alfa-json";
 
 import { Cell } from "./cell";
 import { ColGroup } from "./colgroup";
+import { isElementByName } from "./helpers";
 import { Row } from "./row";
 import { RowGroup } from "./rowgroup";
